lots of clean up
