+++ Tests/ApplicativeRouterTests/ApplicativeRouterTests.swift
@@ -12,9 +12,6 @@ func req(_ method: ApplicativeRouter.Method, _ location: String, _ body: Data? =
 
 class ApplicativeRouterTests: XCTestCase {
   func testRouter() {
-
-    let tmp = Route.home <¢ .get
-
     let routes: [Parser<(), Route>] = [
       // Matches root path: /
       Route.home <¢ .get <*| end,
